data_collector/requirements.txt
flask
grpcio
grpcio-tools
pymongo
requests
python-dotenv